Aggiornamento del file Makefile e dei file correlati per l'implementazione del comando "cd"
